+++ changelog.md
@@ -8,6 +8,7 @@
 * Added community chat to the readme.
 * Made the cookies strict and secure.
 * Removed the redundant _DOMAIN variables.
+* Set un-secure lax cookie when public web frontend url isn't over a secure connection.
 * Fixed bug where the sort choose field item didn't have a hover effect.
 
 ## Released (2020-11-02)